# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for User Service",
        "title": "API Documentation",
        "contact": {},
        "version": "0.2"
    },
    "paths": {
        "/numerical-method/integration/simpson": {
            "post": {
                "description": "Create the simpson data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Simpson"
                ],
                "summary": "Create Simpson",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqSimpson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Simpson"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/integration/simpson/{id}": {
            "get": {
                "description": "Get the simpson data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Simpson"
                ],
                "summary": "Get Simpson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Simpson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Simpson"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/integration/trapezoid": {
            "post": {
                "description": "Create the trapezoid data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trapezoid"
                ],
                "summary": "Create Trapezoid",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validations.ReqTrapezoid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Trapezoid"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/numerical-method/integration/trapezoid/{id}": {
            "get": {
                "description": "Get the trapezoid data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trapezoid"
                ],
                "summary": "Get Trapezoid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trapezoid ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Trapezoid"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Simpson": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interval": {
                    "type": "integer"
                },
                "lower": {
                    "type": "number"
                },
                "upper": {
                    "type": "number"
                }
            }
        },
        "models.Trapezoid": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interval": {
                    "type": "integer"
                },
                "lower": {
                    "type": "number"
                },
                "upper": {
                    "type": "number"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validations.ReqSimpson": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "interval": {
                    "type": "integer"
                },
                "lower": {
                    "type": "number"
                },
                "upper": {
                    "type": "number"
                }
            }
        },
        "validations.ReqTrapezoid": {
            "type": "object",
            "properties": {
                "function": {
                    "type": "string"
                },
                "interval": {
                    "type": "integer"
                },
                "lower": {
                    "type": "number"
                },
                "upper": {
                    "type": "number"
                }
            }
        }
    }
}